Added various sounds and BGM. Fixed some bugs.
Added background music that starts when app launches. Added a clicking
noise for when the user presses start. Added a bullet shoot sound for
when they shoot. Added 2 more textures for levels 4 and 5. Created a
delay when progressing through to the next level and added a placeholder
sound. Fixed some bugs.